requirements.txt
accelerate==1.6.0
aiohappyeyeballs @ file:///croot/aiohappyeyeballs_1734469393482/work
aiohttp @ file:///croot/aiohttp_1734687138658/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
airportsdata==20250224
annotated-types==0.7.0
anyio==4.9.0
anykeystore==0.2
astor==0.8.1
async-timeout @ file:///croot/async-timeout_1732661977001/work
attrs @ file:///croot/attrs_1734533101012/work
autocommand==2.2.2
av==14.3.0
backports.tarfile==1.2.0
blake3==1.0.4
Bottleneck @ file:///croot/bottleneck_1731058641041/work
Brotli @ file:///croot/brotli-split_1736182456865/work
cachetools==5.5.2
certifi @ file:///croot/certifi_1745939216646/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
cloudpickle==3.1.1
compressed-tensors==0.9.3
contourpy @ file:///croot/contourpy_1738160616259/work
cryptacular==1.6.2
cupy-cuda12x==13.4.1
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
datasets @ file:///croot/datasets_1741367990881/work
decord==0.6.0
defusedxml==0.7.1
Deprecated==1.2.18
depyf==0.18.0
dill @ file:///croot/dill_1715094664823/work
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
easydict==1.13
einops==0.8.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
fastrlock==0.8.3
filelock==3.18.0
flash_attn==2.7.4.post1
flashinfer-python==0.2.5
fonttools @ file:///croot/fonttools_1737039080035/work
frozenlist @ file:///croot/frozenlist_1730902802621/work
fsspec @ file:///croot/fsspec_1736274672181/work
gguf==0.16.2
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
hf-xet==1.0.3
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
hupper==1.12.1
idna @ file:///croot/idna_1714398848350/work
importlib_metadata==8.0.0
importlib_resources @ file:///croot/importlib_resources-suite_1720641103994/work
imutils==0.5.4
inflect==7.3.1
interegular==0.3.3
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
Jinja2==3.1.6
jiter==0.9.0
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
lark==1.2.2
line_profiler==4.2.0
llguidance==0.7.18
llvmlite==0.43.0
lm-format-enforcer==0.10.11
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mistral_common==1.5.4
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
modelscope==1.25.0
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict @ file:///croot/multidict_1730905498140/work
multiprocess @ file:///croot/multiprocess_1692294385131/work
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.4
numba==0.60.0
numexpr @ file:///croot/numexpr_1730215937391/work
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
nvitop==1.4.2
oauthlib==3.2.2
openai==1.75.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-semantic-conventions-ai==0.4.3
outlines==0.1.11
outlines_core==0.1.26
packaging @ file:///croot/packaging_1734472117206/work
pandas @ file:///croot/pandas_1732735089971/work/dist/pandas-2.2.3-cp39-cp39-linux_x86_64.whl#sha256=0a51ed2e81ab863e3d00ed6c5049192ce578ecb38fb467d2f9a6585d3c25f666
partial-json-parser==0.2.1.1.post5
PasteDeploy==3.1.0
pbkdf2==1.3
pillow==11.0.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==4.2.2
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.21.1
propcache @ file:///croot/propcache_1744012718189/work
protobuf==4.25.6
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow @ file:///croot/pyarrow_1738229418212/work/python
pycountry==24.6.1
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing @ file:///croot/pyparsing_1731445506121/work
PyQt6==6.7.1
PyQt6_sip @ file:///croot/pyqt-split_1744804475988/work/pyqt_sip
pyramid==2.0.2
pyramid-mailer==0.15.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
python3-openid==3.2.0
pytz @ file:///croot/pytz_1713974312559/work
PyYAML @ file:///croot/pyyaml_1728657952215/work
pyzmq==26.4.0
qwen-vl-utils==0.0.8
ray==2.43.0
referencing==0.36.2
regex @ file:///croot/regex_1736540786412/work
repoze.sendmail==4.4.1
requests @ file:///croot/requests_1730999120400/work
requests-oauthlib==2.0.0
rich==14.0.0
rich-toolkit==0.14.1
rpds-py==0.24.0
safetensors @ file:///croot/safetensors_1741361308036/work
scipy==1.13.1
sentencepiece==0.2.0
shapely @ file:///croot/shapely_1734445883981/work
shellingham==1.5.4
sip @ file:///croot/sip_1738856193618/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
sympy==1.13.1
tdqm==0.0.1
tensorboardX==2.6.2.2
tiktoken==0.9.0
tokenizers @ file:///croot/tokenizers_1741370336077/work
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado @ file:///croot/tornado_1733960490606/work
tqdm @ file:///croot/tqdm_1738943501192/work
transaction==5.0
transformers==4.51.3
translationstring==1.4
triton==3.2.0
typeguard==4.3.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
unicodedata2 @ file:///croot/unicodedata2_1736541023050/work
urllib3 @ file:///croot/urllib3_1737133630106/work
utils==1.0.2
uvicorn==0.34.2
uvloop==0.21.0
velruse==1.1.1
venusian==3.1.1
vllm==0.8.4
watchfiles==1.0.5
WebOb==1.8.9
websockets==15.0.1
wrapt==1.17.2
WTForms==3.2.1
wtforms-recaptcha==0.3.2
xformers==0.0.29.post2
xgrammar==0.1.18
xxhash @ file:///croot/python-xxhash_1737039434400/work
yarl @ file:///croot/yarl_1732546845924/work
zipp @ file:///croot/zipp_1732630741423/work
zope.deprecation==5.1
zope.interface==7.2
zope.sqlalchemy==3.1
